+++ shared/src/commonMain/kotlin/org/hyperskill/app/notification/remote/injection/PushNotificationsComponentImpl.kt
@@ -3,7 +3,7 @@ package org.hyperskill.app.notification.remote.injection
 import org.hyperskill.app.core.injection.AppGraph
 import org.hyperskill.app.notification.remote.data.NotificationTimeRepositoryImpl
 import org.hyperskill.app.notification.remote.domain.interactor.PushNotificationsInteractor
-import org.hyperskill.app.notification.remote.domain.interactor.repository.NotificationTimeRepository
+import org.hyperskill.app.notification.remote.domain.repository.NotificationTimeRepository
 import org.hyperskill.app.notification.remote.remote.NotificationTimeRemoteDataSourceImpl
 
 class PushNotificationsComponentImpl(
